# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance tracking API. All /api/* endpoints require either a session cookie (browser) or an Authorization: Bearer \u003ctoken\u003e header (machine clients).",
        "title": "Cashout API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/web",
    "paths": {
        "/api/analytics/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns total income/expenses and per-category aggregates for a given month.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Monthly category breakdown",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM (defaults to current month)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.MonthlyAnalyticsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/analytics/trend": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Monthly trend over the trailing N months",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of trailing months (1..60, default 12)",
                        "name": "months",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.TrendResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/analytics/year": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Annual breakdown by month and category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "4-digit year (defaults to current year)",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.YearAnalyticsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "web.CategoryBreakdown": {
            "type": "object",
            "properties": {
                "Expense": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.CategoryEntry"
                    }
                },
                "Income": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.CategoryEntry"
                    }
                }
            }
        },
        "web.CategoryEntry": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "pct": {
                    "type": "number"
                }
            }
        },
...
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "web.MonthPoint": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "expense": {
                    "type": "number"
                },
                "income": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "web.MonthlyAnalyticsResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "byCategory": {
                    "$ref": "#/definitions/web.CategoryBreakdown"
                },
                "month": {
                    "type": "string"
                },
                "totalExpenses": {
                    "type": "number"
                },
                "totalIncome": {
                    "type": "number"
                }
            }
        },
...
        "web.TrendResponse": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.MonthPoint"
                    }
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "web.YearAnalyticsResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "byCategory": {
                    "$ref": "#/definitions/web.CategoryBreakdown"
                },
                "byMonth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.YearMonthEntry"
                    }
                },
                "totalExpenses": {
                    "type": "number"
                },
                "totalIncome": {
                    "type": "number"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "web.YearMonthEntry": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "expense": {
                    "type": "number"
                },
                "income": {
                    "type": "number"
                },
                "month": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide the value as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}